Introduce class Serial which incapsulates UART communication, add command which forces stepper to move immediately, add new temporary main which allows direct stepper control over serial line.

+++ src/OS_API/embedded_cout.h
@@ -44,7 +44,7 @@ public:
 	EmbeddedCout& operator<<(const int8_t data);
 	EmbeddedCout& operator<<(const int16_t data);
 	EmbeddedCout& operator<<(const int32_t data);
-	EmbeddedCout& operator<<(const int data);
+	EmbeddedCout& operator<<(const int64_t data);
 
 	EmbeddedCout& operator<<(const float fData);
 };
